+++ src/Database/SqliteConnection.php
@@ -1,9 +1,0 @@
-<?php declare(strict_types=1);
-
-namespace Parable\Orm\Database;
-
-use PDO;
-
-class SqliteConnection extends PDO
-{
-}