+++ tests/Classes/ConnectionClassExtendsPDO.php
@@ -1,0 +1,9 @@
+<?php declare(strict_types=1);
+
+namespace Parable\Orm\Tests\Classes;
+
+use PDO;
+
+class ConnectionClassExtendsPDO extends PDO
+{
+}